PLATFORM-7908 update CurseProfile: fix migration hook
